\name{print.CIFsm}
\alias{print.CIFsm}

\title{
Print CIFsm object
}
\description{
Print the estimated of cumulative incidence functions and point-wise summary statistics. 
}
\usage{
\method{print}{CIFsm}(x, \dots)
}

\arguments{
  \item{x}{
	CIFsm object from function CIFsm
	}
  \item{...}{
  additional arguments to print()
  }
}
\details{
Risk set, cumulative incidence function estimate, standard error and p-value of each group. Point-wise summary statistics estimate, 
standard error and p-value.
}

\seealso{
\code{\link{CIFsm}}
}

\keyword{cumulative incidence function}



