%%File CIFsmry/man/summary.CIFsm.Rd
\name{summary.CIFsm}
\alias{summary.CIFsm}
\title{
Summary of weighted time-integrated summary statistics
}
\description{
Summary statistic estimate, standard error, 95\% confidence interval, p-value depending the method and weight. 
}
\usage{
\method{summary}{CIFsm}(object, ...)
}

\arguments{
  \item{object}{
object of function CIFsm
}
  \item{...}{
  additional arguments to summary()
  }
}

\seealso{
\code{\link{CIFsm}}
}

\keyword{cumulative incidence function}


